@@ -1,6 +1,7 @@
 {
   "name": "andybluntish.com",
   "version": "15.0.0",
+  "type": "module",
   "scripts": {
     "build": "BUILD_ENV=production eleventy",
     "test": "node --test ./tests/",
@@ -26,21 +27,21 @@
     "update-checkin-data": "./scripts/update_checkin_data.js"
   },
   "devDependencies": {
-    "@11ty/eleventy": "^2.0.1",
+    "@11ty/eleventy": "3.0.0-alpha.2",
     "@11ty/eleventy-fetch": "^4.0.0",
     "@11ty/eleventy-navigation": "^0.3.5",
-    "@11ty/eleventy-plugin-webc": "^0.11.1",
+    "@11ty/eleventy-plugin-webc": "^0.11.2",
     "@11ty/eleventy-upgrade-help": "^2.0.5",
     "cheerio": "^1.0.0-rc.12",
-    "clean-css": "^5.3.2",
+    "clean-css": "^5.3.3",
     "dotenv": "^16.3.1",
-    "eslint": "^8.52.0",
-    "eslint-config-prettier": "^9.0.0",
+    "eslint": "^8.56.0",
+    "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-node": "^11.1.0",
-    "eslint-plugin-prettier": "^5.0.1",
+    "eslint-plugin-prettier": "^5.1.0",
     "html-minifier": "^4.0.0",
-    "markdown-it": "^13.0.2",
-    "prettier": "^3.0.3",
+    "markdown-it": "^14.0.0",
+    "prettier": "^3.1.1",
     "purgecss": "^5.0.0"
   }
 }
