@@ -1,6 +1,6 @@
-const { accessSync } = require("fs");
-const EleventyFetch = require("@11ty/eleventy-fetch");
-const {
+import { accessSync } from "fs";
+import EleventyFetch from "@11ty/eleventy-fetch";
+import {
   gravity,
   abv,
   ibu,
@@ -9,9 +9,9 @@ const {
   fermentables,
   hops,
   yeasts,
-} = require("./formatters.js");
+} from "./formatters.js";
 
-class Batch {
+export default class Batch {
   constructor(data = {}) {
     this.data = data;
   }
@@ -197,5 +197,3 @@ async function fetchPage(lastId = null, status = "Archived") {
     return [];
   }
 }
-
-module.exports = Batch;
