@@ -1,7 +1,7 @@
-const cheerio = require("cheerio");
-const { PurgeCSS } = require("purgecss");
-const CleanCSS = require("clean-css");
-const htmlmin = require("html-minifier");
+import cheerio from "cheerio";
+import { PurgeCSS } from "purgecss";
+import CleanCSS from "clean-css";
+import htmlmin from "html-minifier";
 
 async function purge(content) {
   // parse the HTML content
@@ -59,7 +59,7 @@ function minify(content) {
   });
 }
 
-module.exports = async function (content, outputPath) {
+export default async function (content, outputPath) {
   if (!outputPath.endsWith(".html")) {
     return content;
   }
@@ -68,4 +68,4 @@ module.exports = async function (content, outputPath) {
   const minified = minify(purged);
 
   return minified;
-};
+}
