@@ -1,4 +1,4 @@
-module.exports = {
+export default {
   eleventyComputed: {
     title: (data) => `Brew log | ${data.batch.name}`,
     pageStylesheet: (data) => `/beer/batches/${data.batch.slug}`,
