@@ -1,6 +1,6 @@
-const BLANK_VALUE = "-";
+export const BLANK_VALUE = "-";
 
-function uniq(value) {
+export function uniq(value) {
   if (!Array.isArray(value)) {
     return [];
   }
@@ -8,7 +8,7 @@ function uniq(value) {
   return [...new Set(value)];
 }
 
-function abv(value) {
+export function abv(value) {
   let abv = Number(value);
 
   if (!value || Number.isNaN(abv)) {
@@ -24,7 +24,7 @@ function abv(value) {
   }).format(abv);
 }
 
-function ibu(value = 0) {
+export function ibu(value = 0) {
   const ibu = Math.round(value);
 
   if (Number.isNaN(ibu) || !ibu) {
@@ -34,7 +34,7 @@ function ibu(value = 0) {
   }
 }
 
-function ebc(value = 0) {
+export function ebc(value = 0) {
   const ebc = Number(value) * 1.97;
 
   if (Number.isNaN(ebc) || !ebc) {
@@ -44,7 +44,7 @@ function ebc(value = 0) {
   }
 }
 
-function gravity(value = 0) {
+export function gravity(value = 0) {
   const gravity = Number.parseFloat(value);
 
   if (Number.isNaN(gravity) || !gravity) {
@@ -66,7 +66,7 @@ function compareAmounts(a, b) {
   return 0;
 }
 
-function fermentables(data = []) {
+export function fermentables(data = []) {
   const fermentables = Object.values(data)
     .sort(compareAmounts)
     .reverse()
@@ -116,7 +116,7 @@ function normaliseFermentable(value) {
   }
 }
 
-function hops(data = []) {
+export function hops(data = []) {
   const hops = Object.values(data)
     .sort(compareAmounts)
     .reverse()
@@ -138,7 +138,7 @@ function normaliseHop(value) {
   }
 }
 
-function yeasts(data = []) {
+export function yeasts(data = []) {
   const yeasts = Object.values(data)
     .sort(compareAmounts)
     .reverse()
@@ -164,15 +164,3 @@ function normaliseYeast(value) {
       return value;
   }
 }
-
-module.exports = {
-  BLANK_VALUE,
-  uniq,
-  gravity,
-  abv,
-  ibu,
-  ebc,
-  fermentables,
-  hops,
-  yeasts,
-};
